dev: refactor and use test_utils for test
