+++ crates/conformance-test-utils/Cargo.toml
@@ -6,11 +6,13 @@ description = { workspace = true }
 homepage = { workspace = true }
 
 [dependencies]
-clap = { version = "4.2", features = [ "derive" ] }
 starknet = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_with = { workspace = true }
+kakarot_rpc_core = { path = "../core" }
+url = { workspace = true }
+tokio = { workspace = true }
 
 [dev-dependencies]
 cargo-husky = { workspace = true }
